Added UnitTest classes. Database logic not working currently
