+++ UnitTestAPI/UnitTestAPI.csproj
@@ -59,8 +59,12 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="UnitTest1.cs" />
+    <Compile Include="LoginTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="DishesTest.cs" />
+    <Compile Include="RestaurantsTest.cs" />
+    <Compile Include="IngredientsController.cs" />
+    <Compile Include="UserTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
