+++ UnitTestAPI/UserTest.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+
+namespace UnitTestAPI
+{
+    [TestClass]
+    public class UserTest
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
